refactor: split example integration tests by behavior
