+++ tests/integration/src/rust_pipeline/expect/examples/mod.rs
@@ -1,0 +1,3 @@
+mod basic_wallet_package_sizes;
+mod counter_metadata;
+mod storage_metadata;
