+++ tests/integration/src/rust_pipeline/functional/examples/mod.rs
@@ -1,0 +1,4 @@
+mod auth_component_no_auth;
+mod auth_component_rpo_falcon512;
+mod counter_contract_debug_build;
+mod counter_note;
