+++ tests/integration/src/rust_pipeline/property/examples/mod.rs
@@ -1,0 +1,3 @@
+mod collatz;
+mod fibonacci;
+mod is_prime;
